testing a greater diversity of names
* matches now pick up all variants previously listed in issues as failing (see issue #44, issue #97)
* previously a few genus-level matches were matching to `cleanedName` instead of `canonicalName`, resulting in matches to synonymous terms (issue #113)
* as described in issue #114, In commit 8aabbc8 there was an error introduced to all matches where an ifelse() statement was included to created the aligned_name. This has now been fixed.

In implementing this fix, a temporary `aligned_name` field is created that captures the string for the taxon name and this is formatted into the complete `aligned_name` element in a second line of code, removing duplicate code.

The current version has been tested for a list of 225 taxon names that includes at least 2 name strings that are picked up by each of the 50 possible matches.