standardise taxaonomic_dataset values to APC & APNI only
Previously the terms "APC", "APC accepted" and "APC known"  existed as taxonomic dataset values. "APC" (or APNI) are the datasets, while known/accepted refers to the taxonomic status.

The merged terms are still required for generating separate taxonomic resource lists, but not as a value for the variable "taxonomic_dataset"