more changes
I've gone back to requiring that `length(i) <= n_allowed`. The rest of the code I added (loop that starts with `for (j in 1:length(i)) {`) is required if a user ever manually set n_allowed > 1 - the old code would have broken. So for the main APCalign functions (`align_taxa` and `create_taxonomic_update_lookup` n_allowed is already set to 1, but now the fuzzy matching algorithm can be used on its own with a higher limit set.

Tests tweaked again to accommodate this.