Option if fuzzy_match accepted list is empty for a given letter (#218)
Needed to add `if else` loop to `fuzzy_match.R` to only search for fuzzy matches if the subset accepted list (with same first letter) is non-empty. If there were no strings on the accepted list with the same first letter as the input text, warnings were generated.

Test added to check this functionality.